auth-backend: refactor to remove internal hardcoded default session expiration
