--- plugins/auth-backend/src/lib/session/constants.ts
@@ -1,19 +1,0 @@
-/*
- * Copyright 2023 The Backstage Authors
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-// BACKSTAGE_SESSION_EXPIRATION the default session expiration time
-// TODO: find a less hard-coded way to access this, perhaps by reading it from the configuration.
-export const BACKSTAGE_SESSION_EXPIRATION = 3600;
