+++ plugins/auth-backend/src/service/router.test.ts
@@ -15,11 +15,7 @@
  */
 
 import { ConfigReader } from '@backstage/config';
-import {
-  createOriginFilter,
-  getDefaultBackstageTokenExpiryTime,
-} from './router';
-import { BACKSTAGE_SESSION_EXPIRATION } from '../lib/session';
+import { createOriginFilter } from './router';
 
 describe('Auth origin filtering', () => {
   const config = new ConfigReader({
@@ -54,66 +50,5 @@ describe('Auth origin filtering', () => {
   it("Won't explode, valid origin with chars and numbers", () => {
     const origin = 'https://test-test1234.example.net';
     expect(createOriginFilter(config)(origin)).toBeTruthy();
-  });
-});
-
-describe('Test for default backstage token expiry time', () => {
-  it('Will return default backstage session expiration', () => {
-    const config = new ConfigReader({
-      app: {
-        baseUrl: 'http://example.com/extra-path',
-      },
-    });
-    expect(getDefaultBackstageTokenExpiryTime(config)).toBe(
-      BACKSTAGE_SESSION_EXPIRATION,
-    );
-  });
-
-  it('Will return user defined 120 minutes as backstage session expiration', () => {
-    const config = new ConfigReader({
-      app: {
-        baseUrl: 'http://example.com/extra-path',
-      },
-      auth: {
-        backstageTokenExpiration: { minutes: 120 },
-      },
-    });
-    expect(getDefaultBackstageTokenExpiryTime(config)).toBe(7200);
-  });
-
-  it('Will return minimum duration of 10 minutes as backstage session expiration', () => {
-    const config = new ConfigReader({
-      app: {
-        baseUrl: 'http://example.com/extra-path',
-      },
-      auth: {
-        backstageTokenExpiration: { minutes: 2 },
-      },
-    });
-    expect(getDefaultBackstageTokenExpiryTime(config)).toBe(600);
-  });
-
-  it('Will return user configured value as backstage session expiration', () => {
-    const config = new ConfigReader({
-      app: {
-        baseUrl: 'http://example.com/extra-path',
-      },
-      auth: {
-        backstageTokenExpiration: { minutes: 20 },
-      },
-    });
-    expect(getDefaultBackstageTokenExpiryTime(config)).toBe(1200);
-  });
-
-  it('Will return maximum of 24 hour as backstage session expiration if user configured value is more than a day', () => {
-    const config = new ConfigReader({
-      app: {
-        baseUrl: 'http://example.com/extra-path',
-      },
-      auth: {
-        backstageTokenExpiration: { minutes: 1500 },
-      },
-    });
-    expect(getDefaultBackstageTokenExpiryTime(config)).toBe(86400);
   });
 });
